refactor: rationalise conversion functions
