+++ opentelemetry-kotlin-compat/api/opentelemetry-kotlin-compat.api
@@ -17,11 +17,11 @@ public final class io/embrace/opentelemetry/kotlin/creator/ContextCreatorExtKt {
 	public static final fun current (Lio/embrace/opentelemetry/kotlin/creator/ContextCreator;)Lio/embrace/opentelemetry/kotlin/context/Context;
 }
 
-public final class io/embrace/opentelemetry/kotlin/j2k/bridge/context/OtelJavaContextExtKt {
+public final class io/embrace/opentelemetry/kotlin/j2k/bridge/context/OtelJavaContextConversionKt {
 	public static final fun toOtelKotlin (Lio/opentelemetry/context/Context;)Lio/embrace/opentelemetry/kotlin/context/Context;
 }
 
-public final class io/embrace/opentelemetry/kotlin/j2k/logging/export/LogExportConversionKt {
+public final class io/embrace/opentelemetry/kotlin/j2k/logging/export/OtelJavaLogRecordExporterConversionKt {
 	public static final fun toOtelKotlin (Lio/opentelemetry/sdk/logs/export/LogRecordExporter;)Lio/embrace/opentelemetry/kotlin/logging/export/LogRecordExporter;
 }
 
@@ -29,7 +29,7 @@ public final class io/embrace/opentelemetry/kotlin/j2k/tracing/OtelJavaSpanKindC
 	public static final fun toOtelKotlin (Lio/opentelemetry/api/trace/SpanKind;)Lio/embrace/opentelemetry/kotlin/tracing/model/SpanKind;
 }
 
-public final class io/embrace/opentelemetry/kotlin/j2k/tracing/export/SpanExportConversionKt {
+public final class io/embrace/opentelemetry/kotlin/j2k/tracing/export/OtelJavaSpanExporterConversionKt {
 	public static final fun toOtelKotlin (Lio/opentelemetry/sdk/trace/export/SpanExporter;)Lio/embrace/opentelemetry/kotlin/tracing/export/SpanExporter;
 }
 
@@ -54,10 +54,20 @@ public final class io/embrace/opentelemetry/kotlin/k2j/id/TracingIdGeneratorImpl
 	public fun getInvalidTraceId ()Ljava/lang/String;
 }
 
-public final class io/embrace/opentelemetry/kotlin/k2j/tracing/SpanContextOtelJavaConverterKt {
+public final class io/embrace/opentelemetry/kotlin/k2j/tracing/ContextKeyConversionKt {
 	public static final fun toOtelJava (Lio/embrace/opentelemetry/kotlin/context/ContextKey;)Lio/opentelemetry/context/ContextKey;
-	public static final fun toOtelJava (Lio/embrace/opentelemetry/kotlin/tracing/model/SpanContext;)Lio/opentelemetry/api/trace/SpanContext;
+}
+
+public final class io/embrace/opentelemetry/kotlin/k2j/tracing/OtelJavaSpanContextConversionKt {
 	public static final fun toOtelKotlin (Lio/opentelemetry/api/trace/SpanContext;)Lio/embrace/opentelemetry/kotlin/tracing/model/SpanContext;
+}
+
+public final class io/embrace/opentelemetry/kotlin/k2j/tracing/OtelJavaStatusCodeConversionKt {
+	public static final fun toOtelKotlin (Lio/opentelemetry/api/trace/StatusCode;)Lio/embrace/opentelemetry/kotlin/tracing/StatusCode;
+}
+
+public final class io/embrace/opentelemetry/kotlin/k2j/tracing/SpanContextConversionKt {
+	public static final fun toOtelJava (Lio/embrace/opentelemetry/kotlin/tracing/model/SpanContext;)Lio/opentelemetry/api/trace/SpanContext;
 }
 
 public final class io/embrace/opentelemetry/kotlin/k2j/tracing/SpanExtKt {
@@ -66,6 +76,5 @@ public final class io/embrace/opentelemetry/kotlin/k2j/tracing/SpanExtKt {
 
 public final class io/embrace/opentelemetry/kotlin/k2j/tracing/StatusCodeConversionKt {
 	public static final fun toOtelJava (Lio/embrace/opentelemetry/kotlin/tracing/StatusCode;)Lio/opentelemetry/api/trace/StatusCode;
-	public static final fun toOtelKotlin (Lio/opentelemetry/api/trace/StatusCode;)Lio/embrace/opentelemetry/kotlin/tracing/StatusCode;
 }
