+++ opentelemetry-kotlin-compat/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/k2j/tracing/OtelJavaSpanContextConversion.kt
@@ -1,0 +1,8 @@
+package io.embrace.opentelemetry.kotlin.k2j.tracing
+
+import io.embrace.opentelemetry.kotlin.ExperimentalApi
+import io.embrace.opentelemetry.kotlin.aliases.OtelJavaSpanContext
+import io.embrace.opentelemetry.kotlin.tracing.model.SpanContext
+
+@OptIn(ExperimentalApi::class)
+public fun OtelJavaSpanContext.toOtelKotlin(): SpanContext = SpanContextAdapter(this)
