+++ opentelemetry-kotlin-compat/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/k2j/tracing/SpanContextConversion.kt
@@ -1,0 +1,18 @@
+package io.embrace.opentelemetry.kotlin.k2j.tracing
+
+import io.embrace.opentelemetry.kotlin.ExperimentalApi
+import io.embrace.opentelemetry.kotlin.aliases.OtelJavaImmutableSpanContext
+import io.embrace.opentelemetry.kotlin.aliases.OtelJavaSpanContext
+import io.embrace.opentelemetry.kotlin.tracing.model.SpanContext
+
+@OptIn(ExperimentalApi::class)
+public fun SpanContext.toOtelJava(): OtelJavaSpanContext {
+    return OtelJavaImmutableSpanContext.create(
+        traceId,
+        spanId,
+        traceFlags.toOtelJava(),
+        traceState.toOtelJava(),
+        isRemote,
+        false
+    )
+}
